+++ pennylane_lightning/core/src/simulators/lightning_qubit/gates/cpu_kernels/GateImplementationsLM.hpp
@@ -874,26 +874,34 @@ class GateImplementationsLM : public PauliGenerator<GateImplementationsLM> {
               const std::vector<std::size_t> &wires, const bool inverse) {
         using ParamT = PrecisionT;
 
-        const PrecisionT shift = (inverse) ? -1.0 : 1.0;
+        // const PrecisionT shift = (inverse) ? -1.0 : 1.0;
+        constexpr PrecisionT half = 0.5;
+        const Kokkos::complex<PrecisionT> z0{half, (inverse) ? -half : half};
+        const Kokkos::complex<PrecisionT> z1 = Kokkos::conj(z0);
 
-        auto core_function = [shift](std::complex<PrecisionT> *arr,
+        // auto core_function = [shift](std::complex<PrecisionT> *arr,
+        auto core_function = [&z0, &z1](std::complex<PrecisionT> *arr,
                                      const std::size_t i0,
                                      const std::size_t i1) {
             const std::complex<PrecisionT> v0 = arr[i0];
             const std::complex<PrecisionT> v1 = arr[i1];
 
-            const PrecisionT vr_plus = v0.real() + v1.real();
-            const PrecisionT vi_plus = v0.imag() + v1.imag();
+            // const PrecisionT vr_plus = v0.real() + v1.real();
+            // const PrecisionT vi_plus = v0.imag() + v1.imag();
 
-            const PrecisionT vr_minus = (v0.real() - v1.real()) * shift;
-            const PrecisionT vi_minus = (-v0.imag() + v1.imag()) * shift;
+            // const PrecisionT vr_minus = (v0.real() - v1.real()) * shift;
+            // const PrecisionT vi_minus = (-v0.imag() + v1.imag()) * shift;
 
-            arr[i0] = std::complex<PrecisionT>(vr_plus + vi_minus,
-                                               vi_plus + vr_minus) *
-                      PrecisionT(0.5);
-            arr[i1] = std::complex<PrecisionT>(vr_plus - vi_minus,
-                                               vi_plus - vr_minus) *
-                      PrecisionT(0.5);
+            // arr[i0] = std::complex<PrecisionT>(vr_plus + vi_minus,
+            //                                    vi_plus + vr_minus) *
+            //           PrecisionT(0.5);
+            // arr[i1] = std::complex<PrecisionT>(vr_plus - vi_minus,
+            //                                    vi_plus - vr_minus) *
+            //           PrecisionT(0.5);
+
+            arr[i0] = z0 * v0 + z1 * v1;
+            arr[i1] = z1 * v0 + z0 * v1;
+
         };
         if (controlled_wires.empty()) {
             applyNC1<PrecisionT, ParamT, decltype(core_function), false>(